feat : UserDao 추상화, SQLExceptionTranslator 활용
