+++ src/test/java/springbook/user/dao/UserDaoTest.java
@@ -7,11 +7,16 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.support.SQLErrorCodeSQLExceptionTranslator;
+import org.springframework.jdbc.support.SQLExceptionTranslator;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import springbook.user.domain.User;
 
+import javax.sql.DataSource;
 import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.List;
@@ -30,6 +35,9 @@ class UserDaoTest {
     @Autowired
     private UserDao userDao;
     private User wedge;
+
+    @Autowired
+    DataSource dataSource;
 
     @BeforeEach
     void beforeEach() throws SQLException, ClassNotFoundException {
@@ -68,6 +76,21 @@ class UserDaoTest {
         assertThat(expectedWedge).isEqualTo(wedge);
         assertThat(expectedWedge.getName()).isEqualTo(wedge.getName());
         assertThat(expectedWedge.getPassword()).isEqualTo(wedge.getPassword());
+    }
+
+    @Test
+    void sqlExceptionTranslate() {
+        userDao.deleteAll();
+
+        try {
+            userDao.add(USER1);
+            userDao.add(USER1);
+        } catch (DuplicateKeyException ex) {
+            SQLException sqlEx = (SQLException) ex.getRootCause();
+            SQLExceptionTranslator set
+                    = new SQLErrorCodeSQLExceptionTranslator(this.dataSource);
+            assertThat(set.translate(null, null, sqlEx)).isInstanceOf(DataAccessException.class);
+        }
     }
 
     @ParameterizedTest
